r7197 from 4.3.12: SCO client: Fix overflow in memory calculation when >2 GB memory
git-svn-id: svn://svn.code.sf.net/p/xymon/code@7201 44351d6e-118b-4698-b696-ce33095ecaa4

+++ trunk/xymond/client/sco_sv.c
@@ -76,7 +76,7 @@ void handle_sco_sv_client(char *hostname, char *clienttype, enum ostype_t os,
 		char *p;
 
 		memphystotal = memphysfree = 0;
-		memphystotal = (atoi(memsizestr) / 1048576);
+		memphystotal = (atol(memsizestr) / 1048576);
 		if(sscanf(freememstr, "%*s %ld %ld %*d %*d", &memphysfree, &memswapfree) == 2)
 			memphysfree /= 256; /* comes in 4kb pages */
 		else
